+++ cmd/reviewdog/main.go
@@ -280,12 +280,12 @@ func run(r io.Reader, w io.Writer, opt *option) error {
 		// instead of review comment because if it's PR from forked repository,
 		// GitHub token doesn't have write permission due to security concern and
 		// cannot post results via Review API.
-		if cienv.IsInGitHubAction() && cienv.IsGitHubPRFromForkedRepo() {
-			fmt.Fprintln(w, `reviewdog: This is Pull-Request from forked repository.
-GitHub token doesn't have write permission of Review API, so reviewdog will
-report results via logging command [1] and create annotations similar to
+		if cienv.IsInGitHubAction() && cienv.HasReadOnlyPermissionGitHubToken() {
+			fmt.Fprintln(w, `reviewdog: This GitHub token doesn't have write permission of Review API [1], 
+so reviewdog will report results via logging command [2] and create annotations similar to
 github-pr-check reporter as a fallback.
-[1]: https://help.github.com/en/actions/automating-your-workflow-with-github-actions/development-tools-for-github-actions#logging-commands`)
+[1]: https://docs.github.com/en/actions/reference/events-that-trigger-workflows#pull_request_target, 
+[2]: https://help.github.com/en/actions/automating-your-workflow-with-github-actions/development-tools-for-github-actions#logging-commands`)
 			cs = githubutils.NewGitHubActionLogWriter(opt.level)
 		} else {
 			cs = reviewdog.MultiCommentService(gs, cs)